+++ examples/simple/outputs.tf
@@ -20,7 +20,11 @@ output "project_id" {
 }
 
 output "topic_name" {
-  value       = var.topic_name
+  value       = module.pubsub.topic
   description = "The name of the Pub/Sub topic created"
 }
 
+output "topic_labels" {
+  value       = module.pubsub.topic_labels
+  description = "The labels of the Pub/Sub topic created"
+}