+++ examples/simple/variables.tf
@@ -23,3 +23,10 @@ variable "topic_name" {
   type        = string
   description = "The name for the Pub/Sub topic"
 }
+
+variable "topic_labels" {
+  type        = map(string)
+  description = "A map of labels to assign to the Pub/Sub topic"
+  default     = {}
+}
+